pyproject.toml
[project]
name = "mihomo-sync"
version = "0.1.0"
description = "A service to synchronize Mihomo proxy policies with Mosdns DNS rules"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "httpx>=0.23.0",
    "PyYAML>=6.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"